docs-site/requirements.txt
mkdocs-material>=9.5
mkdocs-minify-plugin>=0.8
mkdocs-rss-plugin>=1.6
pillow>=10.0
cairosvg>=2.7
pyyaml>=6.0
